requirements.txt
PyQt6==6.8.0
PyQt6_sip==13.9.1
SQLAlchemy==2.0.37
pytest-qt==4.2.0
